Renaming batch.util to batch.common.
